\name{Alpha-package}
\alias{Alpha-package}
\alias{Alpha}
\docType{package}
\title{
\packageTitle{AlphaPackage}
}
\description{
\packageDescription{Alpha}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Write a function that returns a randomly chosen letter of the alphabet}
\packageIndices{Alpha}
~~Write a function that returns a randomly chosen letter of the alphabet~~
}
\author{ Sharidan Rickmon
\packageAuthor{Sharidan Rickmon}

Maintainer: \packageMaintainer{Sharidan Rickmon}
}
\references{
~~ Write a function that returns a randomly chosen letter of the alphabet~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
