\name{AlphaPackage-package}
\alias{AlphaPackage-package}
\alias{AlphaPackage}
\docType{package}
\title{
\packageTitle{AlphaPackage}
}
\description{
\packageDescription{AlphaPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AlphaPackage}
\packageIndices{AlphaPackage}
~~An overview of how to use the package, including the most important functions~~
}
\author{
\packageAuthor{AlphaPackage}

Maintainer: \packageMaintainer{AlphaPackage}
}
\references{
~~ Write a function that returns a randomly chosen letter of the alphabet~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
