\name{Alphaa-package}
\alias{Alphaa-package}
\alias{Alphaa}
\docType{package}
\title{
\packageTitle{Alphaa}
}
\description{
\packageDescription{Alphaa}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Alphaa}
\packageIndices{Alphaa}
~~Write a function that returns a randomly chosen letter of the
alphabet ~~
}
\author{ Sharidan
\packageAuthor{Alphaa}

Maintainer: \packageMaintainer{Alphaa}
}
\references{
~~ Write a function that returns a randomly chosen letter of the
alphabet ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
